Fix for crash when using this component and not setting any initial text.
If you remove the text from the storyboard that is pre-populating the RichTextEditor, when you tap the RichTextEditor, the app will crash. The reason behind this, is that initially the text views attributed string has no attributes, and consequently the typing attributes has nothing in it, which crashes.

The fix is to wait until the setSelectedTextRange is called when the user first taps in the RichTextEditor to call updateToolbarState.