Fix for left aligning not working after changing the alignment.
If you are changing the alignment for the entire text, sometimes the left alignment won't work again. With the lorem ipsum text in the demo, tap in the center of the screen. Change center align, and then back to left align. The left alignment doesn't work.

Also, I found out that if there is text in the editor already, then we do need to call updateToolbarState.